% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indicators.R
\name{calculate_indicators}
\alias{calculate_indicators}
\title{calculate_indicators}
\usage{
calculate_indicators(data)
}
\arguments{
\item{data}{answers from survey}
}
\value{
a tibble
}
\description{
calculate_indicators
}
