% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_indicators.R
\name{clean_indicators}
\alias{clean_indicators}
\title{clean_indicators}
\usage{
clean_indicators(data)
}
\arguments{
\item{data}{answers from survey}
}
\value{
a cleaned data
}
\description{
clean_indicators
}
