% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_palette_benev.R
\name{create_palette_benev}
\alias{create_palette_benev}
\title{create_palette_benev}
\usage{
create_palette_benev(name_palette = "primaire")
}
\arguments{
\item{name_palette}{choice palette type}
}
\value{
list of palette colors
}
\description{
create_palette_benev
}
