% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_benev.R
\name{scale_fill_benev}
\alias{scale_fill_benev}
\title{scales_fill_benev}
\usage{
scale_fill_benev(name_palette = "green", n_colors = 4)
}
\arguments{
\item{name_palette}{name among benev palette colors}

\item{n_colors}{number of used colors}
}
\value{
aes colors
}
\description{
scales_fill_benev
}
