% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_survey_answers.R
\docType{data}
\name{survey_answers}
\alias{survey_answers}
\title{survey_answers}
\format{
A data frame with 2532 rows and 101 variables:
\describe{
\item{ Identifiant Répondant }{  numeric }
\item{ Type échantillon }{  character }
\item{ Q1 }{  numeric }
\item{ BE1PH }{  numeric }
\item{ BE2PH }{  numeric }
\item{ BE3PS }{  numeric }
\item{ BE4SP }{  numeric }
\item{ BE5PS }{  numeric }
\item{ BE6ENV }{  numeric }
\item{ BE7ENV }{  numeric }
\item{ BE8ENV }{  numeric }
\item{ BE9PH }{  numeric }
\item{ BE10PH }{  numeric }
\item{ BE11ENV }{  numeric }
\item{ BE12ENV }{  numeric }
\item{ BE13ENV }{  numeric }
\item{ BE14PH }{  numeric }
\item{ BE15PH }{  numeric }
\item{ BE16PH }{  numeric }
\item{ BE17PH }{  numeric }
\item{ BE18PS }{  numeric }
\item{ BE19RS }{  numeric }
\item{ BE20RS }{  numeric }
\item{ BE21RS }{  numeric }
\item{ BE22ENV }{  numeric }
\item{ BE23ENV }{  numeric }
\item{ BE24ENV }{  numeric }
\item{ BE25ENV }{  numeric }
\item{ BE26PSR }{  numeric }
\item{ BE27G }{  numeric }
\item{ BE28G }{  numeric }
\item{ CN1 }{  numeric }
\item{ CN2 }{  numeric }
\item{ CN3 }{  numeric }
\item{ CN4 }{  numeric }
\item{ CN5 }{  numeric }
\item{ CN6 }{  numeric }
\item{ CN7 }{  numeric }
\item{ CN8 }{  numeric }
\item{ CN9 }{  numeric }
\item{ CN10 }{  numeric }
\item{ CN11 }{  numeric }
\item{ CN12 }{  numeric }
\item{ CN13 }{  numeric }
\item{ CN14 }{  numeric }
\item{ CN15 }{  numeric }
\item{ CN16 }{  numeric }
\item{ CN17 }{  numeric }
\item{ CN18 }{  numeric }
\item{ CN19 }{  numeric }
\item{ CNDIST }{  numeric }
\item{ EXP1 }{  numeric }
\item{ EXP2 }{  numeric }
\item{ EXP3 }{  numeric }
\item{ EXP4 }{  numeric }
\item{ EXP5 }{  numeric }
\item{ EXP6 }{  numeric }
\item{ EXP7 }{  numeric }
\item{ ACT1 }{  numeric }
\item{ ACT2 }{  numeric }
\item{ ACT3 }{  numeric }
\item{ ACT4 }{  numeric }
\item{ ACT5 }{  numeric }
\item{ ACT6 }{  numeric }
\item{ ACT7 }{  numeric }
\item{ ACT8 }{  numeric }
\item{ ACT9 }{  numeric }
\item{ CP1 }{  numeric }
\item{ CP2 }{  numeric }
\item{ CP3 }{  numeric }
\item{ CP4 }{  numeric }
\item{ CP5 }{  numeric }
\item{ CP6 }{  numeric }
\item{ CP7 }{  numeric }
\item{ CP8 }{  numeric }
\item{ CP9 }{  numeric }
\item{ CP10 }{  numeric }
\item{ CSP }{  numeric }
\item{ TPS-TVAIL }{  numeric }
\item{ TPS-TJT }{  numeric }
\item{ NIV-ET }{  numeric }
\item{ REVENU }{  numeric }
\item{ Sit-FAM }{  numeric }
\item{ ENFANTS }{  numeric }
\item{ FOYER }{  numeric }
\item{ SURFACE }{  numeric }
\item{ TYP-HAB }{  numeric }
\item{ TYP-HAB-Autre }{  character }
\item{ PBLM-SANT }{  character }
\item{ PBLM-SANT-Autre }{  character }
\item{ TABAC }{  numeric }
\item{ ALCOOL }{  numeric }
\item{ DROGUES }{  numeric }
\item{ SEDENT }{  numeric }
\item{ Q31_CSP }{  numeric }
\item{ Q32 }{  character }
\item{ Sexe }{  numeric }
\item{ Année de Naissance }{  numeric }
\item{ Age }{  numeric }
\item{ Code Postal }{  numeric }
\item{ UDA5 - Segmentation géographique }{  numeric }
}
}
\source{
Source
}
\usage{
survey_answers
}
\description{
Description.
}
\keyword{datasets}
