% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_survey_crossref.R
\docType{data}
\name{survey_crossref}
\alias{survey_crossref}
\title{survey_crossref}
\format{
A data frame with  rows and  variables:
\describe{
\item{ SEXE }{  tbl_df,tbl,data.frame }
\item{ AGE }{  tbl_df,tbl,data.frame }
\item{ CSP }{  tbl_df,tbl,data.frame }
\item{ UDA5 }{  tbl_df,tbl,data.frame }
}
}
\source{
Source
}
\usage{
survey_crossref
}
\description{
Cross ref for answers
}
\keyword{datasets}
