% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_survey_questions.R
\docType{data}
\name{survey_questions}
\alias{survey_questions}
\title{survey_questions}
\format{
A data frame with 97 rows and 2 variables:
\describe{
\item{ code }{  character }
\item{ question }{  character }
}
}
\source{
Source
}
\usage{
survey_questions
}
\description{
All questions from the survey
}
\keyword{datasets}
