% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_benev.R
\name{theme_benev}
\alias{theme_benev}
\title{Theme_benev}
\usage{
theme_benev(
  plot_title_family = "Mulish",
  plot_title_color = "#3a7139",
  plot_title_size = 35,
  plot_subtitle_family = "Montserrat",
  plot_subtitle_color = "#1C57BC",
  plot_subtitle_size = 25,
  axis_title_family = "Montserrat",
  axis_title_color = "#C12C6D",
  axis_title_size = 20,
  axis_text_family = "Montserrat",
  axis_text_color = "#C12C6D",
  axis_text_size = 16,
  ...
)
}
\arguments{
\item{plot_title_family}{typo title}

\item{plot_title_color}{color title}

\item{plot_title_size}{size title}

\item{plot_subtitle_family}{typo subtitle}

\item{plot_subtitle_color}{color subtitle}

\item{plot_subtitle_size}{size subtitle}

\item{axis_title_family}{typo axis}

\item{axis_title_color}{color axis}

\item{axis_title_size}{size axis}

\item{axis_text_family}{typo text}

\item{axis_text_color}{color text}

\item{axis_text_size}{size text}

\item{...}{list of other params}
}
\value{
a ggplot theme
}
\description{
Theme_benev
}
